Transitions from Brocfile.js to ember-cli-build.js
See
https://github.com/ember-cli/ember-cli/blob/master/TRANSITION.md#user-co
ntent-brocfile-transition
