+++ ember-cli-build.js
@@ -4,8 +4,11 @@ var EmberApp = require('ember-cli/lib/broccoli/ember-app');
 
 module.exports = function(defaults) {
   var app = new EmberApp(defaults, {
-    // Add options here
+    fingerprint: {
+      prepend: "http://s3.amazonaws.com/frolfr-assets/"
+    }
   });
+  var pickFiles = require('broccoli-static-compiler');
 
   // Use `app.import` to add additional libraries to the generated
   // output files.
@@ -19,6 +22,19 @@ module.exports = function(defaults) {
   // modules that you would like to import into your application
   // please specify an object with the list of modules as keys
   // along with the exports of each module as its value.
+  app.import('bower_components/moment/moment.js');
+  app.import('bower_components/jquery-cookie/jquery.cookie.js');
+  app.import('bower_components/typeahead.js/dist/typeahead.jquery.js');
+  app.import('bower_components/highcharts/highcharts.js');
+  app.import('bower_components/bootstrap-sass-official/assets/javascripts/bootstrap.js');
+  app.import('bower_components/components-font-awesome/css/font-awesome.css');
+  app.import('bower_components/ember-uploader/dist/ember-uploader.js');
+  app.import('bower_components/floatThead/dist/jquery.floatThead.js');
 
-  return app.toTree();
+  var fontAwesome = pickFiles('bower_components/components-font-awesome/fonts', {
+    srcDir: '/',
+    destDir: '/fonts'
+  });
+
+  return app.toTree([fontAwesome]);
 };
